refactor(chat): extract codeBlockFormatter from MessageRenderer
The post-markdown code-block transform (wrap <pre>, language label + copy,
relocate Obsidian's copy button) moves to a pure codeBlockFormatter.ts;
renderContent calls formatCodeBlocks(el). Also drops a now-dead local
runRendererAction helper.

Behavior-preserving (MessageRenderer suites — 108 tests — pass untouched).
MessageRenderer 812 -> 770 LOC.
